Trainer card requires user card first
An authorized user must have their card accepted before it can be
replaced with a training card or proxy card. A training card can
not be used by itself. While a box is authorized with a training
card the display will be purple; while a box is authorized with a
proxy card the display will be orange.