Watchdog optional, no user cache, training cards
The use of a watchdog timer is optional, so we do not automatically
write to /tmp/box_activity every 0.1 second.
Caches of card IDs no longer used. Need to figure out how to make
this work with SQL database.
Use of training cards changed. An authorized trainer must first
use their card, then switch to the training card for this
equipment. Boxes have different colors for proxy, training, and
user cards.
No longer necessary to read the card ID number twice to get a
valid result. This is done in PortalBox instead.